upd
upd

+++ backend/app.py
@@ -72,17 +72,21 @@ def serve_uploaded_file_design(filename):
 def uploaded_news_file(filename):
     return send_from_directory(app.config['UPLOAD_FOLDER_NEWS'], filename)
 
-@app.route('/uploads/mentorsRequest/<path:filename>')
-def uploaded_mentor_file(filename):
-    return send_from_directory(app.config['UPLOAD_FOLDER_MENTORS'], filename, as_attachment=True)
-
-@app.route('/uploads/mentorsRequest/<filename>', methods=['DELETE'])
-def delete_mentor_file(filename):
+@app.route('/uploads/mentorsRequest/<path:filename>', methods=['GET', 'DELETE'])
+def mentors_file(filename):
     file_path = os.path.join(app.config['UPLOAD_FOLDER_MENTORS'], filename)
-    if os.path.exists(file_path):
-        os.remove(file_path)
-        return jsonify({'message': 'Файл удалён'}), 200
-    return jsonify({'error': 'Файл не найден'}), 404
+    
+    if request.method == 'GET':
+        if os.path.exists(file_path):
+            return send_from_directory(app.config['UPLOAD_FOLDER_MENTORS'], filename)
+        return jsonify({'error': 'Файл не найден'}), 404
+
+    if request.method == 'DELETE':
+        if os.path.exists(file_path):
+            os.remove(file_path)
+            return jsonify({'message': 'Файл удалён'}), 200
+        return jsonify({'error': 'Файл не найден'}), 404
+
 
 DATABASE_USERS = 'data/users.db'
 DATABASE_CAMPAIGNS = 'data/campaigns.db'
